加上计时功能，功率改成uint，防止溢出

+++ PDU-CalibrateTool/common/datapacket/datapacket.h
@@ -27,7 +27,7 @@ struct sDataUnit
     ushort cur[PACK_ARRAY_SIZE]; // 电流
     ushort cured[PACK_ARRAY_SIZE]; // 电流
 
-    ushort pow[PACK_ARRAY_SIZE]; // 功率
+    uint pow[PACK_ARRAY_SIZE]; // 功率
     uint ele[PACK_ARRAY_SIZE]; // 电能
 
     uchar pf[PACK_ARRAY_SIZE]; // 功率因数
